move firmware_keys_frame into its own class
